Verification: migrate to Epoxy - create sub packages
