+++ vector/src/main/java/im/vector/riotx/features/crypto/verification/request/VerificationRequestViewModel.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.riotx.features.crypto.verification
+package im.vector.riotx.features.crypto.verification.request
 
 import com.airbnb.mvrx.*
 import com.squareup.inject.assisted.Assisted
@@ -27,6 +27,7 @@ import im.vector.matrix.android.internal.crypto.verification.PendingVerification
 import im.vector.riotx.core.di.HasScreenInjector
 import im.vector.riotx.core.platform.EmptyAction
 import im.vector.riotx.core.platform.VectorViewModel
+import im.vector.riotx.features.crypto.verification.VerificationBottomSheet
 
 data class VerificationRequestViewState(
         val roomId: String? = null,
